+++ src/SpotifyDJApiServer.h
@@ -19,6 +19,7 @@ public:
   using DjResponseCallback = bool (*)(void *context, const String &text, const String &audioUrl, bool &spoken, String &audioType);
   using LanguageProvisionedCallback = void (*)(void *context, const String &languageCode);
   using DeviceCommandCallback = bool (*)(void *context, const DeviceCommand &command, String &message);
+  using DirectPairCallback = void (*)(void *context);
 
   void begin(
       WebServer &server,
@@ -35,7 +36,8 @@ public:
       void *callbackContext = nullptr,
       DjResponseCallback djResponseCallback = nullptr,
       LanguageProvisionedCallback languageProvisionedCallback = nullptr,
-      DeviceCommandCallback deviceCommandCallback = nullptr);
+      DeviceCommandCallback deviceCommandCallback = nullptr,
+      DirectPairCallback directPairCallback = nullptr);
   void loop();
   bool isRunning() const;
 
@@ -66,5 +68,6 @@ private:
   DjResponseCallback djResponseCallback_ = nullptr;
   LanguageProvisionedCallback languageProvisionedCallback_ = nullptr;
   DeviceCommandCallback deviceCommandCallback_ = nullptr;
+  DirectPairCallback directPairCallback_ = nullptr;
   bool running_ = false;
 };